Base Types

+++ Module_4/fake_math.py
@@ -1,4 +1,4 @@
 def divide(first, second):
-    if int(second) == 0:
+    if float(second) == 0.0:
         return 'ERROR'
     return float(first) / float(second)
